refactor: styled components 적용
