+++ src/api.js
@@ -3,11 +3,13 @@ export async function getProducts({
   pageSize = 4,
   orderBy = 'recent',
   keyword = '',
-}) { 
+}) {
   const query = `page=${page}&pageSize=${pageSize}&orderBy=${orderBy}&keyword=${keyword}`;
-  const response = await fetch(`https://panda-market-api.vercel.app/products/?${query}`);
+  const response = await fetch(
+    `https://panda-market-api.vercel.app/products/?${query}`
+  );
   if (!response.ok) {
-    throw new Error("데이터를 불러오는 데 실패했습니다.");
+    throw new Error('데이터를 불러오는 데 실패했습니다.');
   }
   const body = response.json();
   return body;
