+++ src/layouts/HomeLayout.js
@@ -1,0 +1,9 @@
+import { Outlet } from 'react-router-dom';
+
+export default function HomeLayout() {
+  return (
+    <div>
+      <Outlet />
+    </div>
+  );
+}
