+++ src/layouts/LoginLayout.js
@@ -1,0 +1,11 @@
+import { Outlet } from 'react-router-dom';
+import LoginAndSignupTitle from '../components/LoginAndSignupTitle';
+
+export default function LoginLayout() {
+  return (
+    <div>
+      <LoginAndSignupTitle />
+      <Outlet />
+    </div>
+  );
+}
